Add a skeleton for databearDB class to manage sqlite database.
Rudimentary starting point with methods that don't do anything yet
but are commented with what they will do. Anything can still change
but this is a first stab at what a basic database manager will do to
keep management of the configurations and data in the database separate
from the logger itself.